requirements.txt
absl-py==1.2.0
albumentations==1.4.4
annotated-types==0.6.0
antspyx==0.5.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.7
conda-pack==0.7.1
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.16
dominate==2.9.1
efficientnet-pytorch==0.7.1
einops==0.7.0
entrypoints==0.4
executing==1.1.1
fastjsonschema==2.16.2
filelock==3.8.0
fonttools==4.37.4
gevent==24.11.1
google-auth==2.12.0
google-auth-oauthlib==0.4.6
greenlet==3.1.1
grpcio==1.49.1
h5py==3.10.0
huggingface-hub==0.10.1
humanize==4.9.0
idna==3.4
imageio==2.34.0
importlib-metadata==5.0.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
lazy_loader==0.4
lmdb==1.4.1
lxml==5.2.1
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mdurl==0.1.2
MedPy==0.4.0
mistune==2.0.4
munch==4.0.0
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.7
neurite==0.2
nibabel==5.2.1
nltk==3.8.1
notebook==6.4.12
numpy==1.26.4
oauthlib==3.2.1
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
packaging==21.3
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
patsy==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pretrainedmodels==0.7.4
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.7.0
pydantic_core==2.18.1
pydicom==2.4.4
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
pystrum==0.4
python-dateutil==2.8.2
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
qudida==0.0.4
regex==2024.4.16
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.13.0
sckit-learn==0.24.2
segmentation-models-pytorch==0.3.3
Send2Trash==1.8.0
shellingham==1.5.4
SimpleITK==2.2.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.5.1
stack-data==0.5.1
statsmodels==0.14.4
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.16.0
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tifffile==2022.10.10
timm==0.9.2
tinycss2==1.1.1
torch @ file:///remote-home/share/torch-1.11.0%2Bcu115-cp39-cp39-linux_x86_64.whl
torchinfo==1.8.0
torchio==0.19.6
torchsummary==1.5.1
torchvision @ file:///remote-home/share/torchvision-0.12.0%2Bcu115-cp39-cp39-linux_x86_64.whl
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.12
visdom==0.2.4
voxelmorph==0.2
wcwidth==0.2.5
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.16.0
yacs==0.1.8
zipp==3.9.0
zope.event==5.0
zope.interface==7.1.1
